+++ Bundles/UtilText/src/Spryker/Service/UtilText/Model/StringGenerator.php
@@ -16,7 +16,7 @@ class StringGenerator implements StringGeneratorInterface
      */
     public function generateRandomString($length = 32)
     {
-        $tokenLength = $length / 2;
+        $tokenLength = (int)($length / 2);
         $token = bin2hex(random_bytes($tokenLength));
 
         if (strlen($token) !== $length) {
